requirements.txt
# Основные зависимости FastAPI
fastapi~=0.115.12
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
python-jose~=3.4.0
passlib[bcrypt]~=1.7.4
bcrypt==4.0.1
python-dotenv~=1.1.0
alembic
pyttsx3~=2.98
gTTS~=2.5.4
markdown2
pydantic~=2.11.4
pydantic-settings~=2.9.1
starlette~=0.46.2
aiofiles~=24.1.0
email-validator
python-multipart
slowapi~=0.1.9
python-magic~=0.4.27

# Тестовые зависимости
pytest~=8.3.5
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
httpx>=0.24.0
pytest-html>=3.2.0
pytest-benchmark>=4.0.0

# Дополнительные тестовые инструменты
locust>=2.15.1
faker>=19.3.1
factory-boy>=3.3.0

black>=23.1.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

bandit>=1.7.5
safety>=2.3.5